requirements.txt
Django==3.2.7
channels==3.0.4
djangorestframework==3.12.4
drf-yasg==1.20.0
django-cors-headers==3.8.0
mysqlclient==2.0.3
djangochannelsrestframework==0.3.0